requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
xmltodict>=0.12.0
djangorestframework==3.8.2
django-coreapi>=2.3.0
